Reduce API surface of godest package
