+++ pkg/godest/embeds.go
@@ -10,20 +10,16 @@ import (
 
 	"github.com/benbjohnson/hashfs"
 	"github.com/pkg/errors"
-
-	"github.com/sargassum-world/fluitans/pkg/godest/fingerprint"
-	"github.com/sargassum-world/fluitans/pkg/godest/fsutil"
-	tp "github.com/sargassum-world/fluitans/pkg/godest/template"
 )
 
-func ComputeCSPHash(resource []byte) string {
+func computeCSPHash(resource []byte) string {
 	rawHash := sha512.Sum512(resource)
 	encodedHash := base64.StdEncoding.EncodeToString(rawHash[:])
 	return fmt.Sprintf("'sha512-%s'", encodedHash)
 }
 
 func identifyModuleNonpageFiles(templates fs.FS) (map[string][]string, error) {
-	modules, err := fsutil.ListDirectories(templates, tp.FilterModule)
+	modules, err := listDirectories(templates, filterTemplateModule)
 	if err != nil {
 		return nil, errors.Wrap(err, "couldn't list template modules")
 	}
@@ -39,7 +35,7 @@ func identifyModuleNonpageFiles(templates fs.FS) (map[string][]string, error) {
 		if err != nil {
 			return nil, errors.Wrap(err, fmt.Sprintf("couldn't list template module %s", module))
 		}
-		moduleSubfiles, err := fsutil.ListFiles(subfs, tp.FilterNonpage)
+		moduleSubfiles, err := listFiles(subfs, filterNonpageTemplate)
 		moduleFiles[module] = make([]string, len(moduleSubfiles))
 		for i, subfile := range moduleSubfiles {
 			if module == "" {
@@ -60,16 +56,16 @@ func identifyModuleNonpageFiles(templates fs.FS) (map[string][]string, error) {
 func computeAppFingerprint(
 	appAssets, sharedTemplates []string, templates, app fs.FS,
 ) (string, error) {
-	appConcatenated, err := fsutil.ReadConcatenated(appAssets, app)
+	appConcatenated, err := readConcatenated(appAssets, app)
 	if err != nil {
 		return "", errors.Wrap(err, "couldn't load all app assets together for fingerprinting")
 	}
-	sharedConcatenated, err := fsutil.ReadConcatenated(sharedTemplates, templates)
+	sharedConcatenated, err := readConcatenated(sharedTemplates, templates)
 	if err != nil {
 		return "", errors.Wrap(err, "couldn't load all shared templates together for fingerprinting")
 	}
 
-	return fingerprint.Compute(append(appConcatenated, sharedConcatenated...)), nil
+	return computeFingerprint(append(appConcatenated, sharedConcatenated...)), nil
 }
 
 func computePageFingerprints(
@@ -77,7 +73,7 @@ func computePageFingerprints(
 ) (map[string]string, error) {
 	moduleNonpages := make(map[string][]byte)
 	for module, files := range moduleNonpageFiles {
-		loadedNonpages, err := fsutil.ReadConcatenated(files, templates)
+		loadedNonpages, err := readConcatenated(files, templates)
 		if err != nil {
 			return nil, errors.Wrap(err, fmt.Sprintf(
 				"couldn't load non-page template files in template module %s for fingerprinting", module,
@@ -88,7 +84,7 @@ func computePageFingerprints(
 
 	pages := make(map[string][]byte)
 	for _, file := range pageFiles {
-		loadedPage, err := fsutil.ReadFile(file, templates)
+		loadedPage, err := readFile(file, templates)
 		if err != nil {
 			return nil, errors.Wrap(err, fmt.Sprintf(
 				"couldn't load page template %s for fingerprinting", file,
@@ -100,7 +96,7 @@ func computePageFingerprints(
 	pageFingerprints := make(map[string]string)
 	for _, pageFile := range pageFiles {
 		module := path.Dir(pageFile)
-		pageFingerprints[pageFile] = fingerprint.Compute(append(
+		pageFingerprints[pageFile] = computeFingerprint(append(
 			// Each page's fingerprint is computed from the page template itself as well as any non-page
 			// files (e.g. partials) within its module, recursively including all non-page files in
 			// submodules (i.e. subdirectories)
@@ -121,13 +117,13 @@ type Embeds struct {
 	FontsFS     fs.FS
 }
 
-func (e Embeds) ComputeAppFingerprint() (fingerprint string, err error) {
-	appAssetFiles, err := fsutil.ListFiles(e.AppFS, tp.FilterAsset)
+func (e Embeds) computeAppFingerprint() (fingerprint string, err error) {
+	appAssetFiles, err := listFiles(e.AppFS, filterAsset)
 	if err != nil {
 		err = errors.Wrap(err, "couldn't list app assets")
 		return
 	}
-	sharedFiles, err := fsutil.ListFiles(e.TemplatesFS, tp.FilterShared)
+	sharedFiles, err := listFiles(e.TemplatesFS, filterSharedTemplate)
 	if err != nil {
 		err = errors.Wrap(err, "couldn't list shared templates")
 		return
@@ -140,13 +136,13 @@ func (e Embeds) ComputeAppFingerprint() (fingerprint string, err error) {
 	return
 }
 
-func (e Embeds) ComputePageFingerprints() (fingerprints map[string]string, err error) {
+func (e Embeds) computePageFingerprints() (fingerprints map[string]string, err error) {
 	moduleNonpageFiles, err := identifyModuleNonpageFiles(e.TemplatesFS)
 	if err != nil {
 		err = errors.Wrap(err, "couldn't list template module non-page template files")
 		return
 	}
-	pageFiles, err := fsutil.ListFiles(e.TemplatesFS, tp.FilterPage)
+	pageFiles, err := listFiles(e.TemplatesFS, filterPageTemplate)
 	if err != nil {
 		err = errors.Wrap(err, "couldn't list template pages")
 		return
@@ -157,15 +153,6 @@ func (e Embeds) ComputePageFingerprints() (fingerprints map[string]string, err e
 		err = errors.Wrap(err, "couldn't compute fingerprint for page/module templates")
 		return
 	}
-	return
-}
-
-func (e Embeds) NewTemplates(funcs ...template.FuncMap) (r tp.Templates, err error) {
-	pageFiles, err := fsutil.ListFiles(e.TemplatesFS, tp.FilterPage)
-	if err != nil {
-		err = errors.Wrap(err, "couldn't list template pages")
-	}
-	r = tp.NewTemplates(e.TemplatesFS, pageFiles, funcs...)
 	return
 }
 
@@ -195,7 +182,7 @@ type Inlines struct {
 func (i Inlines) ComputeCSSHashesForCSP() (hashes []string) {
 	hashes = make([]string, 0, len(i.CSS))
 	for _, inline := range i.CSS {
-		hashes = append(hashes, ComputeCSPHash([]byte(inline)))
+		hashes = append(hashes, computeCSPHash([]byte(inline)))
 	}
 	return
 }
@@ -203,7 +190,7 @@ func (i Inlines) ComputeCSSHashesForCSP() (hashes []string) {
 func (i Inlines) ComputeJSHashesForCSP() (hashes []string) {
 	hashes = make([]string, 0, len(i.CSS))
 	for _, inline := range i.JS {
-		hashes = append(hashes, ComputeCSPHash([]byte(inline)))
+		hashes = append(hashes, computeCSPHash([]byte(inline)))
 	}
 	return
 }
