+++ pkg/godest/statics.go
@@ -1,17 +1,23 @@
 package godest
 
 import (
+	"fmt"
 	"io/fs"
 	"net/http"
 	"time"
 
 	"github.com/benbjohnson/hashfs"
-
-	"github.com/sargassum-world/fluitans/pkg/godest/httpcache"
 )
 
+func wrapStaticHeader(h http.Handler, age int) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Cache-Control", fmt.Sprintf("public, max-age=%d", age))
+		h.ServeHTTP(w, r)
+	})
+}
+
 func HandleFS(routePrefix string, fsys fs.FS, age time.Duration) http.Handler {
-	return httpcache.WrapStaticHeader(
+	return wrapStaticHeader(
 		http.StripPrefix(routePrefix, http.FileServer(http.FS(fsys))), int(age.Seconds()),
 	)
 }
